requirements.txt
beautifulsoup4==4.12.2
nltk==3.8.1
numpy==1.26.0
openai==0.27.8
pandas==1.4.2
Pillow==10.0.1
plotly==5.11.0
plotly_express==0.4.1
requests==2.28.2
streamlit==1.20.0
streamlit_authenticator==0.2.2
tensorflow==2.14.0
tensorflow_intel
